requirements.txt
langchain
unstructured # Document loading
chromadb # Vector storage
openai # For embeddings
tiktoken  # For embeddings
langchain-community==0.1.0
langchain-openai==0.1.0
python-dotenv==0.21.0
nltk==3.6.7
discord.py==2.0.0
